+++ tests/test_spooker.py
@@ -1,0 +1,60 @@
+import math
+import pathlib
+import pytest
+from pprint import pprint
+import gzip
+import json
+import subprocess
+import tempfile
+
+
+from ccbr_tools.spooker import spooker
+
+
+@pytest.mark.filterwarnings("ignore:UserWarning")
+def test_spooker():
+    out_filename = spooker(
+        pipeline_outdir=pathlib.Path("tests/data/pipeline_run"),
+        pipeline_name="test_pipeline",
+        pipeline_version="0.1.0",
+        pipeline_path="unknown",
+        clean=True,
+        debug="gha",
+    )
+    with gzip.open(out_filename, "rt") as file:
+        spook_dat = json.load(file)
+    # pprint(spook_dat)
+    expected = {
+        "ccbrpipeliner_module": None,
+        "pipeline_name": "test_pipeline",
+        "pipeline_outdir": "tests/data/pipeline_run",
+        "pipeline_path": "unknown",
+        "pipeline_version": "0.1.0",
+        "sample_names": [],
+    }
+    actual = {
+        k: v for k, v in spook_dat["pipeline_metadata"].items() if k in expected.keys()
+    }
+    assert actual == expected
+
+
+def test_spooker_no_outdir():
+    with pytest.raises(FileNotFoundError) as exc_info:
+        spooker(
+            pipeline_outdir=pathlib.Path("tests/data/pipeline_run/does_not_exist"),
+            pipeline_version="0.1.0",
+            pipeline_name="test_pipeline",
+            pipeline_path="unknown",
+            clean=True,
+            debug="gha",
+        )
+    assert str(exc_info.value).startswith("Pipeline output directory does not exist")
+
+
+def test_spooker_cli():
+    assert (
+        "Usage: spooker "
+        in subprocess.run(
+            "spooker --help", shell=True, capture_output=True, text=True
+        ).stdout
+    )